refactor: remove socket.io code
