+++ package.json
@@ -38,7 +38,6 @@
     "@next-auth/mongodb-adapter": "^1.1.3",
     "@next-auth/prisma-adapter": "^1.0.5",
     "@prisma/client": "^4.11.0",
-    "@socket.io/admin-ui": "^0.5.1",
     "@syncedstore/core": "^0.6.0",
     "@syncedstore/react": "^0.6.0",
     "@t3-oss/env-nextjs": "^0.2.1",
@@ -74,8 +73,6 @@
     "react-dom": "^18.3.1",
     "react-joyride": "^2.8.2",
     "react-markdown": "^9.0.1",
-    "socket.io": "^4.7.1",
-    "socket.io-client": "^4.7.1",
     "superjson": "1.12.2",
     "use-immer": "^0.9.0",
     "use-sound": "^4.0.1",
