--- src/hooks/useSocket/index.ts
@@ -1,1 +1,0 @@
-export { useSocket, SocketProvider, socket } from "./useSocket";
