--- src/hooks/useSocket/useSocket.types.ts
@@ -1,15 +1,0 @@
-import { type Socket } from "socket.io-client";
-
-export interface IUseSocketProvider {
-  children: React.ReactNode;
-}
-
-export interface IUseSocketContext {
-  socket: Socket;
-}
-
-// export interface ServerToClientEvents {}
-
-// export interface ClientToServerEvents {}
-
-// export interface InterServerEvents {}
