+++ src/pages/_app.tsx
@@ -10,7 +10,6 @@ import { type AppType } from "next/app";
 import { modals } from "~/components/shared/modals/modalComponents";
 import SEO from "~/config/next-seo.config";
 import { RoomProvider } from "~/hooks/useRoom";
-import { SocketProvider } from "~/hooks/useSocket";
 import { UserProvider } from "~/hooks/useUser";
 import { THEME, cssResolver } from "~/styles/constants";
 
@@ -65,13 +64,11 @@ const MyApp: AppType<{ session: Session | null }> = ({
           <Notifications position="top-center" />
 
           <SessionProvider session={session}>
-            <SocketProvider>
-              <RoomProvider>
-                <UserProvider>
-                  <Component {...pageProps} />
-                </UserProvider>
-              </RoomProvider>
-            </SocketProvider>
+            <RoomProvider>
+              <UserProvider>
+                <Component {...pageProps} />
+              </UserProvider>
+            </RoomProvider>
           </SessionProvider>
         </ModalsProvider>
       </MantineProvider>
