--- src/pages/api/exceptions/NoTeamException.ts
@@ -1,10 +1,0 @@
-import { type Socket } from "socket.io";
-
-export default class NoTeamException {
-  constructor(socket: Socket) {
-    socket.emit("raiseException", {
-      reason: "noTeamException",
-      msg: "This team does not exist anymore"
-    });
-  }
-}
