--- src/pages/api/exceptions/exceptions.types.ts
@@ -1,1 +1,0 @@
-export type TExceptionReason = "noRoomException" | "noTeamException";
