--- src/pages/api/handlers/musicHandlers/index.ts
@@ -1,1 +1,0 @@
-export { musicHandler } from "./musicHandlers";
