--- src/pages/api/handlers/roomHandlers/index.ts
@@ -1,1 +1,0 @@
-export { roomHandler } from "./roomHandlers";
