--- src/pages/api/handlers/teamHandlers/index.ts
@@ -1,1 +1,0 @@
-export { teamHandler } from "./teamHandlers";
