--- src/pages/api/socket.ts
@@ -1,106 +1,0 @@
-import { PrismaClient } from "@prisma/client";
-import { instrument } from "@socket.io/admin-ui";
-import { type NextApiRequest } from "next";
-import { Server, type Socket } from "socket.io";
-import { env } from "~/env.mjs";
-import {
-  type IClientToServerEvents,
-  type IServerSocketData,
-  type IServerToClientEvents,
-  type TNextApiResponse
-} from "~/types/socket.types";
-import Room from "./classes/Room/Room";
-import { roomManager } from "./controllers/RoomManager";
-import { flaggenHandler } from "./handlers/games/flaggenHandlers";
-import { merkenHandler } from "./handlers/games/merkenHandlers";
-import { roomHandler } from "./handlers/roomHandlers";
-import { teamHandler } from "./handlers/teamHandlers";
-import { geheimwoerterHandler } from "./handlers/games/geheimwoerterHandlers";
-import { setHandler } from "./handlers/games/setHandlers";
-import { duSagstHandler } from "./handlers/games/duSagstHandlers";
-import { musicHandler } from "./handlers/musicHandlers";
-
-const prisma = new PrismaClient();
-
-export let io: Server<
-  IClientToServerEvents,
-  IServerToClientEvents,
-  IServerSocketData
->;
-
-export default async function SocketHandler(
-  req: NextApiRequest,
-  res: TNextApiResponse
-) {
-  // means that socket server was already initialized
-  if (res.socket.server.io) {
-    console.log("Socket handler already initialized");
-    res.end();
-    return;
-  }
-
-  console.log("New Socket handler will be initialized");
-
-  io = new Server(
-    // @ts-ignore
-    res.socket.server,
-    {
-      path: "/api/socket/",
-      addTrailingSlash: false,
-      transports: ["websocket", "polling"],
-      cors: {
-        origin: "*",
-        credentials: true,
-        methods: ["GET", "POST"]
-      }
-    }
-  );
-
-  instrument(io, {
-    auth: {
-      type: "basic",
-      username: env.SOCKET_IO_ADMIN_USERNAME,
-      password: env.SOCKET_IO_ADMIN_PASSWORD
-    },
-    mode: "production"
-  });
-
-  res.socket.server.io = io;
-
-  console.log("Initialize Server done. Now starting to fetch rooms from db");
-  // initialize saved rooms for e.g. after a server crash
-  const rooms = await prisma.room.findMany();
-
-  console.log("Rooms found: ", rooms.length);
-
-  rooms.forEach((room) => {
-    const newRoom = new Room(room);
-    roomManager.addRoom(newRoom);
-  });
-
-  const onConnection = (
-    socket: Socket<
-      IClientToServerEvents,
-      IServerToClientEvents,
-      IServerSocketData
-    >
-  ) => {
-    // initialize all handlers
-    roomHandler(io, socket);
-    teamHandler(io, socket);
-
-    // music and sound effects
-    musicHandler(io, socket);
-
-    // handler for all games
-    flaggenHandler(io, socket);
-    merkenHandler(io, socket);
-    geheimwoerterHandler(io, socket);
-    setHandler(io, socket);
-    duSagstHandler(io, socket);
-  };
-
-  io.on("connection", onConnection);
-
-  res.end();
-}
