+++ src/types/user.types.ts
@@ -1,0 +1,6 @@
+import type { SafedUser } from "~/server/api/routers/users";
+
+export type TUserReduced = Pick<
+  SafedUser,
+  "id" | "name" | "email" | "image" | "role" | "username"
+>;
